requirements.txt
mkdocs-material
pyyaml
